prodbi_datalake_sync/requirements.txt
azure-functions
azure-storage-file-datalake
